Adjusted types and changed mokito syntax

+++ src/apis/DataBaseHandler.java
@@ -2,8 +2,8 @@ package apis;
 public interface DataBaseHandler{
 
 
-    public String storeData(long l);
-    //store long l into something? maybe a file
+    public String storeData(int l);
+    //store int l into something? maybe a file
     //send message to successfully store data
 
     public long sendData(String s);
